requirements.txt
# Specify the version constraint for setuptools
setuptools<58

# Basic libraries for handling VCF files
pyvcf

# Other useful libraries for data handling and analysis
pandas
numpy
scipy

# Optional: Biopython for additional bioinformatics tools
biopython

# kipoiseq
kipoiseq

# pyfaidx
pyfaidx

# sgkit
sgkit

# pyliftover
pyliftover

# tokenizers
tokenizers

# mavehgvs
mavehgvs

# BEND
webdataset

# Olmo
boto3
